pyproject.toml
[tool.poetry]
name = "ggongsul-api"
version = "0.1.0"
description = "api server for ggongsul"
authors = ["jaegeon <zezaeoh@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.1.3"
djangorestframework = "^3.12.2"
uWSGI = "^2.0.19"
mysqlclient = "^2.0.1"
requests = "^2.24.0"
django-storages = "^1.10.1"
Pillow = "^8.0.1"
boto3 = "^1.16.23"
slack-sdk = "^3.1.0"
djangorestframework-simplejwt = "^4.6.0"
django-ckeditor = "^6.0.0"
django-filter = "^2.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
